[JIFFY-293] Make locale settings configurable in Jiffy
