+++ lib/jiffy.dart
@@ -5,4 +5,7 @@ library jiffy;
 
 export 'src/enums/unit.dart';
 export 'src/enums/start_of_week.dart';
+export 'src/locale/locale.dart';
+export 'src/locale/ordinals.dart';
+export 'src/locale/relative_date_time.dart';
 export 'src/jiffy.dart';
