--- lib/src/locale/locales/tr_locale.dart
@@ -1,52 +1,0 @@
-import '../../enums/start_of_week.dart';
-import '../locale.dart';
-import '../relative_date_time.dart';
-
-class TrLocale extends Locale {
-  @override
-  String code() => 'tr';
-
-  @override
-  List<String> ordinals() => List.from(['', '', '', ''], growable: false);
-
-  @override
-  StartOfWeek startOfWeek() => StartOfWeek.monday;
-
-  @override
-  RelativeDateTime relativeDateTime() => TrRelativeTime();
-}
-
-class TrRelativeTime extends RelativeDateTime {
-  @override
-  String prefixAgo() => '';
-  @override
-  String prefixFromNow() => '';
-  @override
-  String suffixAgo() => 'önce';
-  @override
-  String suffixFromNow() => 'içinde';
-  @override
-  String lessThanOneMinute(int seconds) => 'birkaç saniye';
-  @override
-  String aboutAMinute(int minutes) => 'bir dakika';
-  @override
-  String minutes(int minutes) => '$minutes dakika';
-  @override
-  String aboutAnHour(int minutes) => 'bir saat';
-  @override
-  String hours(int hours) => '$hours saat';
-  @override
-  String aDay(int hours) => 'bir gün';
-  @override
-  String days(int days) => '$days gün';
-  @override
-  String aboutAMonth(int days) => 'bir ay';
-  @override
-  String months(int months) => '$months ay';
-  @override
-  String aboutAYear(int year) => 'bir yıl';
-  @override
-  String years(int years) => '$years yıl';
-  @override
-  String wordSeparator() => ' ';
-}
