+++ pubspec.yaml
@@ -11,6 +11,7 @@ environment:
 
 dependencies:
   intl: ">=0.19.0 <0.21.0"
+  quiver: "^3.2.2"
 
 dev_dependencies:
   lints: ">=2.0.1 <4.0.0"
